Display pages with dynamic heights
There is an issue where pages with only 4, 5, or 6 cards will not align
the second row of cards properly due to the height of the page being
more than "necessary".

When printing, this is not an issue because the height is determined by
the content. This commit fixes the issue by making the page height
always be determined by the content, though limited to actual A4 size.